fix(#341): display fly maneuverability on CharacterSheet
