+++ assets/translations/en/charsheet/speed.json5
@@ -3,9 +3,16 @@
     base: 'Base',
     reduced: 'Enc.',
     fly: 'Fly',
-    maneuverability: 'Maneuverability',
     swim: 'Swim',
     climb: 'Climb',
     burrow: 'Burr.',
     misc: 'Misc',
+    maneuverability: {
+        h: 'Maneuverability',
+        perfect: 'Perfect',
+        good: 'Good',
+        average: 'Average',
+        poor: 'Poor',
+        clumsy: 'Clumsy',
+    },
 }
