+++ src/components/CharacterSheet/MainPage/Speed/base.tsx
@@ -1,0 +1,8 @@
+import { SpeedUnit } from '../../../../types/Units';
+
+interface TargetUnit {
+    unit: SpeedUnit;
+    suffix: string;
+}
+
+export type { TargetUnit };
